Fixed 64-bit portability problems so that it now works on Linux/x86_64.
git-svn-id: file:///home/tjaden/{repo}/redstone/trunk@508 ec8ac367-a008-0410-9901-8f3136df4969

+++ libnet/lib/drivers/serial.c
@@ -217,7 +217,7 @@ static struct packet *get_next_packet (struct hl_port *hl_port, int *to_slot)
 
     while (qsize >= PACKET_SIZE) {
 
-        if (marker == PACKET_MARKER) {
+        if ((unsigned int)marker == PACKET_MARKER) {
 
  	    int p = qtail;
